Move locality directives into common
