+++ app/components/appMainModule/localityPageModule/localityPageController.js
@@ -1,7 +1,5 @@
 'use strict';
 
-var _ = require('lodash');
-
 function localityPageController($scope, $state, locality) {
   $scope.locality = locality;
 }
