+++ app/components/common/localityListing/localityListingController.js
