+++ app/components/common/localityListing/localityListingDirective.js
